Also listen for writes to error_fd when polling serial connection
The serial connection would add a file handler to listen for
any input on its port. However this would only handle normal input,
and there was not an equivalent handler to handle any input on
the error file descriptor. As a consequence, it was possible to
get into a situation where gdb was only listening for normal
input (and not errors), while the other end was blocked trying to
write errors.

This fixes the problem but adding an additional handler to
handle error inputs. Any errors are simply forwarded to gdb's stderr.

gdb/ChangeLog:

	* ser-base.c (reschedule): Add file handlers for scb->error_fd
	in places where file handlers are installed for scb->fd.
	(error_fd_event): Added function to forward error input from
	the serial line to gdb's stderr.
	(ser_base_async): Delete file handler for scb->error_fd.